Compile perf tests with the C++ compiler
This lets us build the binaries in a portable fashion w/o having to worry
about how to link with the C++ runtime.

+++ perf/local_lat.cpp
@@ -25,7 +25,7 @@ int main (int argc, char *argv [])
 {
     const char *bind_to;
     int roundtrip_count;
-    int message_size;
+    size_t message_size;
     void *ctx;
     void *s;
     int rc;
